[UPUP-1084]: add FileSampleService
