+++ Sources/MUDKit/Services/MUDKitService.swift
@@ -13,7 +13,6 @@ public struct MUDKitService {
     
     /// Creates sample files in Documents and Temporary directories to test FileSystemService features.
     public static func createSampleFilesForFileSystemServiceDemo() {
-        let fileSystemService = FileSystemService()
-        fileSystemService.createSampleFiles()
+        FileSampleService.createSampleFiles()
     }
 }
